+++ bioformats/fasta.py
@@ -4,9 +4,12 @@
 # Copyright (C) 2015 by Gaik Tamazian
 # gaik (dot) tamazian (at) gmail (dot) com
 
-import pyfaidx
 import random
+
+import pyfaidx
 from builtins import range  # pylint:disable=redefined-builtin
+
+from .bed import Reader, Writer
 from .exception import BioformatsError
 
 
@@ -123,3 +126,64 @@ class Reorder(object):
                                               '}'.format(i))
                 else:
                     seq_writer.write(i, str(seq_reader[i]))
+
+
+class NeighborhoodFilter(object):
+    """
+    The class implements routines for filtering features from a FASTA
+    or VCF file by the presence of N's in their neighborhood.
+    """
+    def __init__(self, fasta_filename, flank_len=100):
+        """
+        Create a Neighborhood filter object.
+
+        :param fasta_filename: a name of a FASTA file
+        :param flank_len: the length of a region to be checked
+        :type fasta_filename: str
+        :type flank_len: int
+        """
+        self.__len = flank_len
+        self.__fasta = pyfaidx.Fasta(fasta_filename)
+
+    def check_seq(self, seq_id, start_pos, end_pos):
+        """
+        Check if the specified region contains N's. Positions are
+        zero-based and half-opened as in the BED format,
+
+        :param seq_id: a sequence ID
+        :param start_pos: a start position in bp
+        :param end_pos: a start position in bp
+        :type seq_id: str
+        :type start_pos: int
+        :type end_pos: int
+        :return: True if the specified region contains no N's,
+            False otherwise
+        :rtype: bool
+        """
+        seq = str(self.__fasta[seq_id][start_pos:end_pos])
+        return 'N' not in seq
+
+    def filter_bed(self, input_filename, output_filename):
+        """
+        Filter routines from the specified BED file and output the
+        filtered ones to the specified output file.
+
+        :param input_filename: a name of a BED file containing
+            features to be filtered
+        :param output_filename: the output BED file name
+        :type input_filename: str
+        :type output_filename: str
+        """
+        with open(input_filename) as input_file:
+            bed_reader = Reader(input_file)
+            with Writer(output_filename) as output_file:
+                for record in bed_reader.records():
+                    seq = record.seq
+                    down_start = max(0, record.start - self.__len)
+                    down_end = record.start
+                    up_start = record.end
+                    up_end = min(len(self.__fasta[record.seq]),
+                                 record.end + self.__len, )
+                    if self.check_seq(seq, down_start, down_end) and \
+                            self.check_seq(seq, up_start, up_end):
+                        output_file.write(record)